+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### NEW FEATURES
 * Added Box-Cox transformation to `effect.size`
+* Added an argument to flip axes in `plot.gm.prcomp`
 
 ### BUG FIXES
 * Fixed plotting parameter arguments for phylo.par in `summary.gm.prcomp`.